requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.1.0
pandas==1.3.5
numpy==1.21.6
fair-esm==2.0.0
scikit-learn===1.0.2
torch==1.12.0+cpu
joblib==1.2.0
gunicorn
